refactor(git): split diff.go into diff/fetch/mergebase files
